made some imporovements on the footer, made it responsive
