+++ client/src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import "./App.css";
 //import LoginPage from "./Components/Pages/login_page/LoginPage";
 //import SignUp from "./Components/Pages/sign-up page/SignUp";
-import Navbar from "./Components/Static-components/Navbar/Navbar";
+// import Navbar from "./Components/Static-components/Navbar/Navbar";
 import Footer from "./Components/Pages/Footer/footer";
 
 function App() {
