+++ client/src/Components/Pages/Footer/footer.css
@@ -1,44 +1,35 @@
-.footer {  
-
-    height: 350px;
-    color: white;
-    background-color: #267A89;
-    padding-top: 5em;
-    padding-bottom: 20em;
-    width: 100%;
-    font-family: 'Manrope';
-    font-style: normal;
-    font-weight: 200;
-    font-size: 18px;
-    line-height: 180%;
-    display: flex;
-    padding-left: 180px;
-    padding-right: 180px;
-    
-    }
-    
-    
-    a:hover {
-        text-decoration: underline;
-    }
-    
- 
-    .footer div{
-
-        text-align: left;
-
-    }
-
-.image{ 
-        flex-grow: 0.2;
-    }
-    .row1{ 
-        flex-grow: 0.15;
-    }
-    .row2{ 
-        flex-grow: 0.15;
-    }
-    .row3{ 
-        flex-grow: 0.15;
-    }
-   
+.footer {
+  display: flex;
+  flex-flow: column wrap;
+  gap: 1rem;
+  padding: 1rem;
+  color: white;
+  background-color: var(--nypd-blue);
+  font: 200 1rem var(--ff-manrope);
+  line-height: 2rem;
+}
+.grid-container {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 1.5rem;
+}
+.grid-container h4 {
+  padding-block-end: 1rem;
+}
+@media screen and (min-width: 500px) {
+  .grid-container {
+    grid-template-columns: repeat(4, 1fr);
+    padding-inline: 2rem;
+  }
+}
+a:hover {
+  text-decoration: underline;
+}
+.image {
+  max-width: 65%;
+  cursor: pointer;
+}
+.image:hover,
+.image:focus {
+  opacity: 0.8;
+}
