+++ client/src/index.css
@@ -7,14 +7,12 @@
 *::before,
 *::after {
   box-sizing: border-box;
-}
-* {
   margin: 0;
+  padding: 0;
 }
 html,
 body {
   height: 100%;
-  
 }
 body {
   line-height: 1.5;
@@ -34,7 +32,9 @@ textarea,
 select {
   font: inherit;
 }
-ul, ol, a {
+ul,
+ol,
+a {
   list-style: none;
   text-decoration: none;
   color: inherit;
